Address some of the review comments from Bin Feng.
