+++ core/server/common/src/main/java/alluxio/cli/validation/SecureHdfsValidationTask.java
@@ -13,9 +13,6 @@ package alluxio.cli.validation;
 
 import alluxio.PropertyKey;
 
-import java.io.BufferedReader;
-import java.io.InputStreamReader;
-import java.io.IOException;
 import java.util.regex.Matcher;
 import java.util.regex.Pattern;
 
@@ -23,9 +20,15 @@ import java.util.regex.Pattern;
  * Task for validating security configurations.
  */
 public class SecureHdfsValidationTask implements ValidationTask {
-  private static final String LINE_SEPARATOR = System.getProperty("line.separator").toString();
   private static final String ALLUXIO_BIN = "./bin/alluxio";
   private static final String GETCONF_CMD = "getConf";
+  /**
+   * Regular expression to parse principal used by Alluxio to connect to secure
+   * HDFS.
+   *
+   * @see <a href="https://web.mit.edu/kerberos/krb5-1.5/krb5-1.5.4/doc/krb5-user/What-is-a-Kerberos-Principal_003f.html">Kerberos documentation</a>
+   * for more details.
+   */
   private static final Pattern PRINCIPAL_PATTERN =
       Pattern.compile("(?<primary>[\\w][\\w-]*\\$?)(/(?<instance>[\\w]+))?(@(?<realm>[\\w]+))?");
 
@@ -49,14 +52,14 @@ public class SecureHdfsValidationTask implements ValidationTask {
     String instance;
     String realm;
     if (mProcess.equals("master")) {
-      principal = getResultFromProcess(new String[] {ALLUXIO_BIN, GETCONF_CMD,
+      principal = Utils.getResultFromProcess(new String[] {ALLUXIO_BIN, GETCONF_CMD,
           PropertyKey.MASTER_PRINCIPAL.getName()}).getOutput();
-      keytab = getResultFromProcess(new String[] {ALLUXIO_BIN, GETCONF_CMD,
+      keytab = Utils.getResultFromProcess(new String[] {ALLUXIO_BIN, GETCONF_CMD,
           PropertyKey.MASTER_KEYTAB_KEY_FILE.getName()}).getOutput();
     } else if (mProcess.equals("worker")) {
-      principal = getResultFromProcess(new String[] {ALLUXIO_BIN, GETCONF_CMD,
+      principal = Utils.getResultFromProcess(new String[] {ALLUXIO_BIN, GETCONF_CMD,
           PropertyKey.WORKER_PRINCIPAL.getName()}).getOutput();
-      keytab = getResultFromProcess(new String[] {ALLUXIO_BIN, GETCONF_CMD,
+      keytab = Utils.getResultFromProcess(new String[] {ALLUXIO_BIN, GETCONF_CMD,
           PropertyKey.WORKER_KEYTAB_FILE.getName()}).getOutput();
     }
     Matcher matchPrincipal = PRINCIPAL_PATTERN.matcher(principal);
@@ -70,7 +73,7 @@ public class SecureHdfsValidationTask implements ValidationTask {
 
     // Login with principal and keytab
     int exitVal =
-        getResultFromProcess(new String[] {"kinit", "-kt", keytab, principal}).getExitValue();
+        Utils.getResultFromProcess(new String[] {"kinit", "-kt", keytab, principal}).getExitValue();
     if (exitVal != 0) {
       System.err.format("Kerberos login failed for %s with keytab %s with exit value %d.%n",
           principal, keytab, exitVal);
@@ -79,63 +82,5 @@ public class SecureHdfsValidationTask implements ValidationTask {
       return false;
     }
     return true;
-  }
-
-  private ProcessExecutionResult getResultFromProcess(String[] args) {
-    try {
-      Process process = Runtime.getRuntime().exec(args);
-      StringBuilder outputSb = new StringBuilder();
-      try (BufferedReader processOutputReader = new BufferedReader(
-          new InputStreamReader(process.getInputStream()))) {
-        String line;
-        while ((line = processOutputReader.readLine()) != null) {
-          outputSb.append(line);
-          outputSb.append(LINE_SEPARATOR);
-        }
-      }
-      StringBuilder errorSb = new StringBuilder();
-      try (BufferedReader processErrorReader = new BufferedReader(
-          new InputStreamReader(process.getErrorStream()))) {
-        String line;
-        while ((line = processErrorReader.readLine()) != null) {
-          errorSb.append(line);
-          errorSb.append(LINE_SEPARATOR);
-        }
-      }
-      process.waitFor();
-      return new ProcessExecutionResult(process.exitValue(), outputSb.toString().trim(),
-          errorSb.toString().trim());
-    } catch (IOException e) {
-      System.err.println("Failed to execute command.");
-      return new ProcessExecutionResult(1, "", "");
-    } catch (InterruptedException e) {
-      System.err.println("Interrupted.");
-      Thread.currentThread().interrupt();
-      return new ProcessExecutionResult(1, "", "");
-    }
-  }
-
-  static class ProcessExecutionResult {
-    private final int mExitValue;
-    private final String mOutput;
-    private final String mError;
-
-    public ProcessExecutionResult(int val, String output, String error) {
-      mExitValue = val;
-      mOutput = output;
-      mError = error;
-    }
-
-    public int getExitValue() {
-      return mExitValue;
-    }
-
-    public String getOutput() {
-      return mOutput;
-    }
-
-    public String getError() {
-      return mError;
-    }
   }
 }
